feat(codesandbox): support preview both `Block` and `Page`
