+++ mockServer/package.json
@@ -25,7 +25,7 @@
     "lint": "eslint --fix ."
   },
   "dependencies": {
-    "@opentiny/tiny-engine-dsl-vue": "^1.0.6",
+    "@opentiny/tiny-engine-dsl-vue": "workspace:*",
     "@seald-io/nedb": "^4.0.2",
     "fs-extra": "^11.1.1",
     "glob": "^10.3.4",
