+++ packages/design-core/src/preview/src/preview/usePreview.js
@@ -14,7 +14,8 @@ import { provide, inject } from 'vue'
 import { ReplStore } from '@vue/repl'
 import vueJsx from '@vue/babel-plugin-jsx'
 import { transformSync } from '@babel/core'
-import { genSFCWithDefaultPlugin, parseRequiredBlocks } from '@opentiny/tiny-engine-dsl-vue'
+import { parseRequiredBlocks } from '@opentiny/tiny-engine-controller/js/block'
+import { genSFCWithDefaultPlugin } from '@opentiny/tiny-engine-dsl-vue'
 import importMap from './importMap'
 import srcFiles from './srcFiles'
 import generateMetaFiles, { processAppJsCode } from './generate'
