+++ packages/toolbars/codesandbox/package.json
@@ -27,7 +27,9 @@
     "@opentiny/tiny-engine-canvas": "workspace:*",
     "@opentiny/tiny-engine-common": "workspace:*",
     "@opentiny/tiny-engine-controller": "workspace:*",
-    "@opentiny/tiny-engine-utils": "workspace:*"
+    "@opentiny/tiny-engine-dsl-vue": "workspace:*",
+    "@opentiny/tiny-engine-utils": "workspace:*",
+    "codesandbox": "^2.2.3"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.2.3",
