+++ packages/toolbars/codesandbox/src/Main.vue
@@ -15,10 +15,15 @@
 </template>
 
 <script>
+import { reactive } from 'vue'
 import { Popover } from '@opentiny/vue'
-import { previewPage, previewBlock } from '@opentiny/tiny-engine-controller/js/preview'
+import { previewBlock } from '@opentiny/tiny-engine-controller/js/preview'
+import getPreGenerateInfo from '@opentiny/tiny-engine-controller/js/generate-files'
 import { getGlobalConfig, useBlock, useCanvas, useLayout, useNotify } from '@opentiny/tiny-engine-controller'
 import { constants } from '@opentiny/tiny-engine-utils'
+import { generateApp } from '@opentiny/tiny-engine-dsl-vue'
+import { getParameters } from 'codesandbox/lib/api/define'
+import { codesandboxFiles } from './codesandboxFiles'
 
 const { PREVIEW_SANDBOX } = constants
 
@@ -33,10 +38,14 @@ export default {
     }
   },
   setup() {
-    const { isBlock, getCurrentPage, canvasApi } = useCanvas()
+    const state = reactive({
+      generating: false
+    })
+
+    const { isBlock, canvasApi } = useCanvas()
     const { getCurrentBlock } = useBlock()
 
-    const preview = () => {
+    const preview = async () => {
       if (useLayout().isEmptyPage()) {
         useNotify({
           type: 'warning',
@@ -60,10 +69,55 @@ export default {
         params.pageInfo.name = block?.label
         previewBlock(params, PREVIEW_SANDBOX.CodeSandbox)
       } else {
-        const page = getCurrentPage()
-        params.id = page?.id
-        params.pageInfo.name = page?.name
-        previewPage(params, PREVIEW_SANDBOX.CodeSandbox)
+        if (state.generating) {
+          useNotify({ type: 'info', title: '代码生成中, 请稍后...' })
+          return
+        } else {
+          useNotify({ type: 'info', title: '代码生成中...' })
+          state.generating = true
+        }
+
+        try {
+          const instance = generateApp()
+          const fileRes = await getPreGenerateInfo(instance, PREVIEW_SANDBOX.CodeSandbox)
+
+          const files = {}
+          fileRes.forEach((file) => {
+            // 使用 pnpm 包管理
+            if (file.filePath === 'README.md') {
+              files[file.filePath] = { content: file.fileContent.replace(/npm /g, 'pnpm ') }
+            } else {
+              files[file.filePath] = { content: file.fileContent }
+            }
+          })
+          Object.assign(files, codesandboxFiles)
+
+          const parameters = getParameters({ files, template: 'vue-cli' })
+          const form = document.createElement('form')
+          form.method = 'POST'
+          form.action = 'https://codesandbox.io/api/v1/sandboxes/define'
+          form.target = '_blank'
+          const parametersInput = document.createElement('input')
+          parametersInput.name = 'parameters'
+          parametersInput.value = parameters
+          const queryInput = document.createElement('input')
+          queryInput.name = 'query'
+          queryInput.value = 'module=/src/App.vue'
+          const environmentInput = document.createElement('input')
+          environmentInput.name = 'environment'
+          environmentInput.value = 'server'
+          form.appendChild(parametersInput)
+          form.appendChild(queryInput)
+          form.appendChild(environmentInput)
+          document.body.append(form)
+          form.submit()
+          document.body.removeChild(form)
+        } catch (error) {
+          // eslint-disable-next-line no-console
+          console.error(error)
+          useNotify({ type: 'error', title: '代码生成失败', message: error?.message || error })
+          state.generating = false
+        }
       }
     }
