+++ packages/vue-generator/src/index.js
@@ -21,5 +21,3 @@ export {
   genSFCWithDefaultPlugin,
   generateSFCFile
 } from './generator'
-
-export { parseRequiredBlocks } from './utils/parseRequiredBlocks'
